app/pyproject.toml
[tool.poetry]
name = "fast-api-template"
version = "0.1.0"
description = ""
authors = ["timofejkazakov <t.kazakov@agora.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
loguru = "^0.7.2"
sqlalchemy = "^2.0.25"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
dependency-injector = "^4.41.0"
flask-security = "^3.0.0"
python-jose = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
